mindsettler-backend/requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python-http-client==3.3.7
sendgrid==6.12.5
sqlparse==0.5.4
Werkzeug==3.1.4
whitenoise==6.6.0